# openapi.yaml
openapi: "3.1.0"
info:
  title: Qovira API
  version: "0.1"
  description: >-
    The Qovira server API. All endpoints are under /api/v1 and speak application/json. Error responses use application/problem+json (RFC 9457) with a `code` extension field that clients branch on.

    Authentication: session cookie __Host-qovira_session (set by POST /auth/login). CSRF: POST/PATCH/DELETE require a CSRF-Token header echoed from the qovira_csrf cookie (double-submit pattern).
servers:
  - url: /api/v1
    description: Qovira v1 API (same-origin)

# ---------------------------------------------------------------------------
# Reusable schemas
# ---------------------------------------------------------------------------
components:
  schemas:
...
    Problem:
      type: object
      required: [type, title, status, detail, code, requestId]
      properties:
        type:
          type: string
          format: uri
          description: Stable URL identifying the problem kind. Always "https://qovira.ai/errors/" + code.
          example: "https://qovira.ai/errors/not_found"
        title:
          type: string
          description: Short, stable, human-readable summary of the problem type.
          example: "Not found"
        status:
          type: integer
          description: HTTP status code; matches the response status line.
          example: 404
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence.
          example: "The requested resource was not found."
        code:
          type: string
          description: Snake_case machine slug; the primary client branching key.
          example: "not_found"
        requestId:
          type: string
          description: Correlation ID tying this response to server logs. Always present; sentinel "unknown" if none in context.
          example: "a1b2c3d4e5f6"
        errors:
          type: array
          description: Per-field validation errors. Present on 422 only.
          items:
            $ref: "#/components/schemas/FieldError"
    FieldError:
      type: object
      required: [pointer, detail]
      properties:
        pointer:
          type: string
          description: RFC 6901 JSON Pointer to the offending field.
          example: "/email"
        detail:
          type: string
          description: Human-readable explanation of why the value is invalid.
          example: "must be a valid email address"
...
    Conversation:
      type: object
      required: [id, preview, createdAt, updatedAt]
      properties:
        id:
          type: string
          description: Opaque conversation identifier.
        preview:
          type: string
          description: First user message truncated to ~80 characters. Empty string when no user message exists yet.
        createdAt:
          type: string
          format: date-time
          description: RFC 3339 UTC timestamp of when the conversation was created.
        updatedAt:
          type: string
          format: date-time
          description: RFC 3339 UTC timestamp of the most recent activity.
    HistoryMessage:
      type: object
      required: [id, role, content, createdAt, abandoned]
      properties:
        id:
          type: string
          description: Opaque message identifier.
        role:
          type: string
          description: Message role — one of user, assistant, tool, system.
        content:
          type: string
          description: Message content. For tool messages this is a JSON string.
        createdAt:
          type: string
          format: date-time
        toolCalls:
          type: array
          description: Present on assistant messages that triggered tool calls. Absent otherwise.
          items:
            type: object
            additionalProperties: true
        toolCallId:
          type: string
          description: Present on tool-result messages. References the callId of the originating tool call. Absent otherwise.
        finishReason:
          type: string
          description: Model finish reason (stop, tool_calls, length, etc.). Present on completed assistant messages. Absent otherwise.
        abandoned:
          type: boolean
          description: True when this assistant message was abandoned because its pending tool confirmation expired. Such messages are inert and must not be treated as outstanding work.
    ConversationHistory:
      type: object
      required: [id, createdAt, updatedAt, messages]
      properties:
        id:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/HistoryMessage"
    MessageResponse:
      type: object
      required: [id, conversationId, role, content, createdAt]
      properties:
        id:
          type: string
        conversationId:
          type: string
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    Pagination:
      type: object
      required: [nextCursor, hasMore]
      properties:
        nextCursor:
          type: string
          nullable: true
          description: Opaque cursor for the next page. null when no more pages.
        hasMore:
          type: boolean
  responses:
    ProblemUnauthorized:
      description: Authentication required (unauthenticated or invalid_credentials).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ProblemForbidden:
      description: CSRF validation failed (csrf_failed).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ProblemNotFound:
      description: Resource not found.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ProblemConflict:
      description: Conflict (e.g. confirmation already resolved or expired).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ProblemUnprocessableEntity:
      description: Validation error (validation_error, validation_failed, invalid_decision).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ProblemBadRequest:
      description: Malformed body or unknown query parameter.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
paths:
...
  /conversations:
    get:
      operationId: listConversations
      summary: List conversations (cursor-paginated, most-recently-active first)
      description: Returns a cursor-paginated list of the caller's conversations ordered by most-recently-active first (updated_at DESC, id DESC). Each item includes an opaque preview derived from the first user message in the conversation, truncated to ~80 characters. Unknown query parameters are rejected with 400 unknown_query_param.
      tags: [conversations]
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
          description: Opaque pagination cursor from a previous response's nextCursor field.
        - name: limit
          in: query
          schema:
            type: integer
            default: 25
            minimum: 1
            maximum: 100
          description: Page size (default 25, max 100).
      responses:
        "200":
          description: Paginated list of conversations.
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Conversation"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "400":
          $ref: "#/components/responses/ProblemBadRequest"
        "401":
          $ref: "#/components/responses/ProblemUnauthorized"
  /conversations/{id}:
    get:
      operationId: getConversation
      summary: Get conversation history
      description: Returns the full chronological message history for the conversation. Returns 404 when the conversation does not exist or belongs to another user.
      tags: [conversations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Conversation ID.
      responses:
        "200":
          description: Conversation with full message history.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationHistory"
        "401":
          $ref: "#/components/responses/ProblemUnauthorized"
        "404":
          $ref: "#/components/responses/ProblemNotFound"
  /conversations/{id}/messages:
    post:
      operationId: postConversationMessage
      summary: Send a message to a conversation
      description: Persists the user message and returns 202 immediately with the persisted message body. The AI response streams over the /events SSE channel tagged by conversationId — the turn outlives this request.
      tags: [conversations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Conversation ID.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
      responses:
        "202":
          description: Message persisted. AI response will stream over SSE.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/ProblemUnauthorized"
        "403":
          $ref: "#/components/responses/ProblemForbidden"
        "404":
          $ref: "#/components/responses/ProblemNotFound"
  /conversations/{id}/confirmations/{callId}:
    post:
      operationId: postConversationConfirmation
      summary: Resolve a pending tool confirmation
      description: Approve or deny a pending tool call confirmation. Returns 202 with no body. The harness re-enters the turn asynchronously via SSE.
      tags: [conversations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Conversation ID.
        - name: callId
          in: path
          required: true
          schema:
            type: string
          description: Pending confirmation call ID.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision:
                  type: string
                  enum: [approve, deny]
      responses:
        "202":
          description: Decision recorded. No body.
        "401":
          $ref: "#/components/responses/ProblemUnauthorized"
        "403":
          $ref: "#/components/responses/ProblemForbidden"
        "404":
          $ref: "#/components/responses/ProblemNotFound"
        "409":
          $ref: "#/components/responses/ProblemConflict"
        "422":
          $ref: "#/components/responses/ProblemUnprocessableEntity"

  # ---------- Reminders ----------
...